backend/booking-service/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
motor==2.5.1
pymongo==3.12.0
python-dotenv==1.0.0
pydantic==1.10.7
boto3==1.26.115
httpx==0.24.0
python-multipart==0.0.6
bcrypt==4.0.1
pytest==7.3.1
pytest-asyncio==0.21.0
dnspython==2.4.2